# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Drug API",
        "description": "مستندات API جستجوی دارو",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/drugs/search": {
            "get": {
                "tags": [
                    "Drugs"
                ],
                "summary": "جستجوی دارو",
                "description": "جستجوی دارو بر اساس نام با امکان فیلتر و صفحه‌بندی",
                "operationId": "searchDrugs",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "نام دارو برای جستجو",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 100,
                            "minLength": 2
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "تعداد آیتم در هر صفحه",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "with_relations",
                        "in": "query",
                        "description": "دریافت اطلاعات مرتبط (گروه‌های دارویی و درمانی)",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "goroh_daroei_cod",
                        "in": "query",
                        "description": "فیلتر بر اساس گروه دارویی",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "goroh_darmani_cod",
                        "in": "query",
                        "description": "فیلتر بر اساس گروه درمانی",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "موفق - لیست داروها",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "نتایج جستجو"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "استامینوفن"
                                                    },
                                                    "brand": {
                                                        "type": "string",
                                                        "example": "تیلنول"
                                                    },
                                                    "dosage": {
                                                        "type": "string",
                                                        "example": "500mg"
                                                    },
                                                    "goroh_daroei": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "خطای درخواست - پارامترهای نامعتبر",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "خطا در اعتبارسنجی داده‌ها"
                                        },
                                        "errors": {
                                            "properties": {
                                                "q": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "فیلد q الزامی است"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "خطای داخلی سرور",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "خطا در جستجو"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "پیغام خطای فنی"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}